+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Tags:
 - **New Feature**
   - Multiple handlers can be registered for the same action.
   - Async action handlers are sent a `dispatchId` as part of the payload. The dispatch id for a `begin` dispatch is equal to the id of the corresponding `success` or `error` dispatch. This enables fine-grained optimistic updates.
+  - `Store#registerMatcher(matcher, handler)` calls a handler when the matcher function returns true for a dispatched payload. The handler is passed the payload.
 - **Breaking Change**
   - The arguments passed to async action handlers have been changed. The final argument passed to any action handler is the raw payload that is sent through the dispatcher. For example, "begin" handlers receive the payload as its sole argument.
 - **Internal**